Cleanup tests.
